+++ sys/arch/evbppc/nintendo/dev/si.h
@@ -143,6 +143,7 @@ struct siio_send {
 	uint32_t	outsize;	/* number of bytes for out buffer */
 	void		*in;		/* buffer to store response */
 	void		*out;		/* buffer to send out to ext device */
+	long		delay;		/* delay the transaction (microsec) */
 };
 
 static inline int
@@ -162,6 +163,12 @@ __si_send(struct si_softc *sc, struct siio_send *data)
 
 	WR4(sc, SISR, SISR_ERROR_MASK(chan));
 	sisr = RD4(sc, SISR);
+
+
+	/* 
+	 * libogc uses interrupts for non-blocking. i think this is ok for now
+	 */
+	delay(data->delay);
 
 	AWAIT_SICOMCSR(sc);
 	WR4(sc, SICOMCSR,
